w3c-reports/12092014/gtk/runned-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-remove-source-expected.txt



FAIL pointer updates (removing source elements) assert_equals: error events on b expected 1 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-remove-source.html:33:18
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
